fix(#3875): 清理重复测试文件和空壳占位测试
- 删除14个重复 *_pytest.py 文件，保留更完整版本并重命名
- 删除3个无价值的 demo 测试（仅 assertTrue(True)）
- 删除5个空壳集成测试（仅 assert True 占位 + @skip）
