--- tests/unit/database/drivers/test_base_driver_pytest.py
@@ -1,321 +1,0 @@
-"""
-性能: 220MB RSS, 2.04s, 17 tests [PASS]
-基础驱动测试 - 使用Pytest最佳实践重构。
-
-测试DatabaseDriverBase基类的核心功能。
-"""
-
-import pytest
-from unittest.mock import Mock, MagicMock, patch
-from datetime import datetime
-import time
-
-try:
-    from ginkgo.data.drivers.base_driver import DatabaseDriverBase
-    from ginkgo.libs import GLOG
-    from ginkgo.libs.core.logger import GinkgoLogger
-except ImportError:
-    DatabaseDriverBase = None
-    GLOG = None
-    GinkgoLogger = None
-
-
-def _make_concrete_driver(uri="test://user:pass@host:port/db",
-                          streaming_uri="test://user:pass@host:port/db"):
-    """创建实现了所有抽象方法的ConcreteDriver类."""
-    class ConcreteDriver(DatabaseDriverBase):
-        def __init__(self, user, pwd, host, port, db, **kwargs):
-            super().__init__("TestDriver")
-            self._user = user
-            self._pwd = pwd
-            self._host = host
-            self._port = port
-            self._db = db
-            self._echo = kwargs.get("echo", False)
-            self._connect_timeout = kwargs.get("connect_timeout", 10)
-            self._read_timeout = kwargs.get("read_timeout", 10)
-
-        def _get_uri(self):
-            return uri
-
-        def _health_check_query(self):
-            return "SELECT 1"
-
-        def _create_engine(self):
-            return None
-
-        def _create_streaming_engine(self):
-            return None
-
-        def _get_streaming_uri(self):
-            return streaming_uri
-
-    return ConcreteDriver
-
-
-@pytest.mark.database
-class TestDatabaseDriverBaseUnit:
-    """DatabaseDriverBase单元测试."""
-
-    @pytest.fixture
-    def mock_logger(self):
-        """创建mock logger."""
-        logger = Mock()
-        logger.INFO = Mock()
-        logger.WARNING = Mock()
-        logger.ERROR = Mock()
-        logger.DEBUG = Mock()
-        return logger
-
-    @pytest.fixture
-    def base_driver(self, mock_logger):
-        """创建基础驱动实例."""
-        if DatabaseDriverBase is None:
-            pytest.skip("DatabaseDriverBase not available")
-
-        ConcreteDriver = _make_concrete_driver()
-        driver = ConcreteDriver(
-            user="test_user",
-            pwd="test_password",
-            host="localhost",
-            port="1234",
-            db="test_db"
-        )
-
-        # 替换logger
-        driver.loggers = [mock_logger]
-
-        return driver
-
-    @pytest.mark.integration
-    def test_base_driver_initialization(self, base_driver):
-        """测试基础驱动初始化."""
-        assert base_driver.driver_name == "TestDriver"
-        assert base_driver._db_type == "testdriver"
-        assert base_driver._user == "test_user"
-        assert base_driver._pwd == "test_password"
-        assert base_driver._host == "localhost"
-        assert base_driver._port == "1234"
-        assert base_driver._db == "test_db"
-
-    @pytest.mark.integration
-    def test_base_driver_default_parameters(self, base_driver):
-        """测试默认参数设置."""
-        assert base_driver._echo is False
-        assert base_driver._connect_timeout == 10
-        assert base_driver._read_timeout == 10
-
-    @pytest.mark.integration
-    def test_base_driver_connection_stats_initialization(self, base_driver):
-        """测试连接统计初始化."""
-        stats = base_driver._connection_stats
-
-        assert stats is not None
-        assert "connections_created" in stats
-        assert "connections_closed" in stats
-        assert "active_connections" in stats
-        assert "created_at" in stats
-        assert "last_health_check" in stats
-        assert "health_check_failures" in stats
-        # streaming stats
-        assert "streaming_connections_created" in stats
-        assert "active_streaming_connections" in stats
-
-    @pytest.mark.integration
-    def test_base_driver_get_connection_stats(self, base_driver):
-        """测试获取连接统计."""
-        stats = base_driver.get_connection_stats()
-
-        assert isinstance(stats, dict)
-        assert "driver_name" in stats
-        assert "connections_created" in stats
-        assert "connections_closed" in stats
-        assert "active_connections" in stats
-        assert "uptime" in stats
-        assert "connection_efficiency" in stats
-
-    @pytest.mark.integration
-    def test_base_driver_log_method(self, base_driver, mock_logger):
-        """测试日志方法."""
-        # 测试不同级别的日志
-        base_driver.log("INFO", "Test info message")
-        assert mock_logger.INFO.called
-
-        base_driver.log("WARNING", "Test warning message")
-        assert mock_logger.WARNING.called
-
-        base_driver.log("ERROR", "Test error message")
-        assert mock_logger.ERROR.called
-
-
-@pytest.mark.database
-class TestDatabaseDriverBaseConnectionManagement:
-    """测试DatabaseDriverBase连接管理功能."""
-
-    @pytest.fixture
-    def base_driver(self):
-        """创建基础驱动实例."""
-        if DatabaseDriverBase is None:
-            pytest.skip("DatabaseDriverBase not available")
-
-        ConcreteDriver = _make_concrete_driver()
-        return ConcreteDriver(
-            user="test_user",
-            pwd="test_password",
-            host="localhost",
-            port="1234",
-            db="test_db"
-        )
-
-    @pytest.mark.integration
-    def test_base_driver_health_check_default(self, base_driver):
-        """测试默认健康检查实现."""
-        # 默认情况下，健康检查应该返回True或False
-        try:
-            result = base_driver.health_check()
-            assert isinstance(result, bool)
-        except (NotImplementedError, RuntimeError):
-            # 如果没有实现引擎或健康检查失败，这是预期的
-            pass
-
-    @pytest.mark.integration
-    def test_base_driver_session_property(self, base_driver):
-        """测试会话属性."""
-        try:
-            session = base_driver.session
-            # 如果有session_factory，应该返回会话
-            # 如果没有，可能会抛出异常
-            pass
-        except (AttributeError, NotImplementedError, RuntimeError):
-            # 如果没有实现session_factory或初始化失败，这是预期的
-            pass
-
-    @pytest.mark.integration
-    def test_base_driver_engine_property(self, base_driver):
-        """测试引擎属性."""
-        try:
-            engine = base_driver.engine
-            # 如果有engine，应该返回引擎对象
-            # 如果没有，可能会抛出异常
-            pass
-        except (AttributeError, NotImplementedError, RuntimeError):
-            # 如果没有实现engine或初始化失败，这是预期的
-            pass
-
-
-@pytest.mark.database
-class TestDatabaseDriverBaseLoggerManagement:
-    """测试DatabaseDriverBase日志管理."""
-
-    @pytest.fixture
-    def mock_logger(self):
-        """创建mock logger."""
-        logger = Mock()
-        logger.INFO = Mock()
-        logger.WARNING = Mock()
-        logger.ERROR = Mock()
-        logger.DEBUG = Mock()
-        return logger
-
-    @pytest.fixture
-    def base_driver(self, mock_logger):
-        """创建基础驱动实例."""
-        if DatabaseDriverBase is None:
-            pytest.skip("DatabaseDriverBase not available")
-
-        ConcreteDriver = _make_concrete_driver()
-        driver = ConcreteDriver(
-            user="test_user",
-            pwd="test_password",
-            host="localhost",
-            port="1234",
-            db="test_db"
-        )
-
-        # 替换logger
-        driver.loggers = [mock_logger]
-
-        return driver
-
-    @pytest.mark.integration
-    def test_base_driver_logger_initialization(self, base_driver):
-        """测试日志初始化."""
-        assert isinstance(base_driver.loggers, list)
-        assert len(base_driver.loggers) > 0
-
-    @pytest.mark.integration
-    @pytest.mark.parametrize("level,message", [
-        ("INFO", "Test info message"),
-        ("WARNING", "Test warning message"),
-        ("ERROR", "Test error message"),
-        ("DEBUG", "Test debug message"),
-    ])
-    def test_base_driver_log_levels(self, base_driver, mock_logger, level, message):
-        """测试不同日志级别."""
-        base_driver.log(level, message)
-
-        # 验证相应的日志方法被调用
-        level_upper = level.upper()
-        assert getattr(mock_logger, level_upper).called
-
-
-@pytest.mark.database
-class TestDatabaseDriverBaseThreadSafety:
-    """测试DatabaseDriverBase线程安全性."""
-
-    @pytest.fixture
-    def base_driver(self):
-        """创建基础驱动实例."""
-        if DatabaseDriverBase is None:
-            pytest.skip("DatabaseDriverBase not available")
-
-        ConcreteDriver = _make_concrete_driver()
-        return ConcreteDriver(
-            user="test_user",
-            pwd="test_password",
-            host="localhost",
-            port="1234",
-            db="test_db"
-        )
-
-    @pytest.mark.integration
-    def test_base_driver_lock_initialization(self, base_driver):
-        """测试锁初始化."""
-        assert base_driver._lock is not None
-        # 验证是一个线程锁对象
-        import threading
-        assert isinstance(base_driver._lock, type(threading.Lock()))
-
-
-@pytest.mark.database
-@pytest.mark.parametrize("driver_type,host,port", [
-    ("clickhouse", "localhost", "8123"),
-    ("mysql", "localhost", "3306"),
-    ("mongodb", "localhost", "27017"),
-])
-class TestDatabaseDriverBaseParametrized:
-    """参数化的基础驱动测试."""
-
-    @pytest.mark.integration
-    def test_base_driver_port_types(self, driver_type, host, port):
-        """测试不同端口类型."""
-        if DatabaseDriverBase is None:
-            pytest.skip("DatabaseDriverBase not available")
-
-        ConcreteDriver = _make_concrete_driver(
-            uri=f"{driver_type}://user:pass@host:port/db",
-            streaming_uri=f"{driver_type}://user:pass@host:port/db",
-        )
-
-        # 测试字符串端口
-        driver = ConcreteDriver(
-            user="test_user",
-            pwd="test_password",
-            host=host,
-            port=str(port),
-            db="test_db"
-        )
-
-        assert driver._port == str(port)
-        assert driver._host == host
-        assert driver._db == "test_db"
