--- tests/unit/lab/test_selector_base_pytest.py
@@ -1,377 +1,0 @@
-"""
-性能: 156MB RSS, 0.87s, 0 tests [PASS]
-Selector基础测试 - 使用Pytest最佳实践重构。
-
-测试BaseSelector、FixedSelector等选择器类的功能。
-"""
-
-import pytest
-from datetime import datetime
-from typing import List, Dict, Any
-from unittest.mock import Mock, patch
-
-try:
-    from ginkgo.backtest.selectors.base_selector import BaseSelector
-    from ginkgo.backtest.selectors.fixed_selector import FixedSelector
-    from ginkgo.backtest.selectors.popularity_selector import PopularitySelector
-    from ginkgo.backtest.selectors.cn_all_selector import CNAllSelector
-    from ginkgo.backtest.portfolios.base_portfolio import BasePortfolio
-except ImportError:
-    BaseSelector = None
-    FixedSelector = None
-    PopularitySelector = None
-    CNAllSelector = None
-    BasePortfolio = None
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestBaseSelectorConstruction:
-    """测试BaseSelector类的构造和初始化."""
-
-    @pytest.mark.unit
-    def test_baseselector_init(self):
-        """测试BaseSelector默认初始化."""
-        if BaseSelector is None:
-            pytest.skip("BaseSelector not available")
-
-        s = BaseSelector()
-
-        # 验证基本属性
-        assert s is not None
-        assert getattr(s, 'portfolio', None) is not None or s.portfolio is None
-        assert getattr(s, 'name', None) is not None
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestBaseSelectorProperties:
-    """测试BaseSelector类的属性管理."""
-
-    @pytest.fixture
-    def mock_portfolio(self):
-        """Mock Portfolio实例."""
-        portfolio = Mock(spec=BasePortfolio)
-        portfolio.uuid = "test_portfolio_uuid"
-        return portfolio
-
-    @pytest.mark.unit
-    def test_baseselector_bind_portfolio(self, mock_portfolio):
-        """测试BaseSelector绑定组合."""
-        if BaseSelector is None:
-            pytest.skip("BaseSelector not available")
-
-        s = BaseSelector()
-        assert s.portfolio is None
-
-        s.bind_portfolio(mock_portfolio)
-
-        assert s.portfolio is not None
-        assert s.portfolio == mock_portfolio
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestBaseSelectorBusinessLogic:
-    """测试BaseSelector类的业务逻辑."""
-
-    @pytest.fixture
-    def mock_portfolio(self):
-        """Mock Portfolio实例."""
-        portfolio = Mock(spec=BasePortfolio)
-        portfolio.uuid = "test_portfolio_uuid"
-        return portfolio
-
-    @pytest.mark.unit
-    def test_baseselector_pick_method_exists(self, mock_portfolio):
-        """测试BaseSelector的pick方法存在."""
-        if BaseSelector is None:
-            pytest.skip("BaseSelector not available")
-
-        s = BaseSelector()
-        s.bind_portfolio(mock_portfolio)
-
-        # 验证pick方法存在
-        assert callable(getattr(s, 'pick', None))
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestFixedSelectorConstruction:
-    """测试FixedSelector类的构造和初始化."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("name,codes", [
-        ("test_fixed_selector", ["halo", "nihao"]),
-        ("test_selector_2", ["000001.SZ", "600000.SH"]),
-        ("test_selector_3", ["AAPL", "GOOGL", "MSFT"]),
-    ])
-    def test_fixedselector_init(self, name, codes):
-        """测试FixedSelector初始化."""
-        if FixedSelector is None:
-            pytest.skip("FixedSelector not available")
-
-        s = FixedSelector(name, codes)
-
-        # 验证基本属性
-        assert s is not None
-        assert s.name == name
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestFixedSelectorSelection:
-    """测试FixedSelector的选择逻辑."""
-
-    @pytest.fixture
-    def mock_portfolio(self):
-        """Mock Portfolio实例."""
-        portfolio = Mock(spec=BasePortfolio)
-        portfolio.uuid = "test_portfolio_uuid"
-        return portfolio
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("name,codes,expected_count", [
-        ("test_selector", ["halo", "nihao"], 2),
-        ("test_selector_2", ["000001.SZ"], 1),
-        ("test_selector_3", [], 0),
-    ])
-    def test_fixedselector_pick(self, mock_portfolio, name, codes, expected_count):
-        """测试FixedSelector选择股票."""
-        if FixedSelector is None:
-            pytest.skip("FixedSelector not available")
-
-        s = FixedSelector(name, codes)
-        s._now = "2021-01-01"
-        s.bind_portfolio(mock_portfolio)
-
-        # 执行选择
-        result = s.pick()
-
-        # 验证结果
-        assert len(result) == expected_count
-        if expected_count > 0:
-            assert all(code in result for code in codes)
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestPopularitySelectorConstruction:
-    """测试PopularitySelector类的构造和初始化."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("name,limit", [
-        ("test_popularity_selector", 10),
-        ("test_selector_2", 20),
-        ("test_selector_3", -10),
-    ])
-    def test_popularityselector_init(self, name, limit):
-        """测试PopularitySelector初始化."""
-        if PopularitySelector is None:
-            pytest.skip("PopularitySelector not available")
-
-        s = PopularitySelector(name, limit)
-
-        # 验证基本属性
-        assert s is not None
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestPopularitySelectorSelection:
-    """测试PopularitySelector的选择逻辑."""
-
-    @pytest.fixture
-    def mock_portfolio(self):
-        """Mock Portfolio实例."""
-        portfolio = Mock(spec=BasePortfolio)
-        portfolio.uuid = "test_portfolio_uuid"
-        return portfolio
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("name,limit", [
-        ("test_selector", 10),
-        ("test_selector_2", -10),
-    ])
-    def test_popularityselector_pick(self, mock_portfolio, name, limit):
-        """测试PopularitySelector选择股票."""
-        if PopularitySelector is None:
-            pytest.skip("PopularitySelector not available")
-
-        s = PopularitySelector(name, limit)
-        s._now = "2020-01-01"
-        s.bind_portfolio(mock_portfolio)
-
-        # 执行选择
-        result = s.pick()
-
-        # 验证结果（PopularitySelector的具体行为取决于实现）
-        assert isinstance(result, list)
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestCNAllSelectorConstruction:
-    """测试CNAllSelector类的构造和初始化."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("name", [
-        "test",
-        "cn_all_selector",
-        "china_market_selector",
-    ])
-    def test_cnallselector_init(self, name):
-        """测试CNAllSelector初始化."""
-        if CNAllSelector is None:
-            pytest.skip("CNAllSelector not available")
-
-        s = CNAllSelector(name)
-
-        # 验证基本属性
-        assert s is not None
-        assert s.name == name
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestCNAllSelectorSelection:
-    """测试CNAllSelector的选择逻辑."""
-
-    @pytest.fixture
-    def mock_portfolio(self):
-        """Mock Portfolio实例."""
-        portfolio = Mock(spec=BasePortfolio)
-        portfolio.uuid = "test_portfolio_uuid"
-        return portfolio
-
-    @pytest.mark.unit
-    def test_cnallselector_pick(self, mock_portfolio):
-        """测试CNAllSelector选择股票."""
-        if CNAllSelector is None:
-            pytest.skip("CNAllSelector not available")
-
-        s = CNAllSelector("test")
-        s.bind_portfolio(mock_portfolio)
-
-        # 执行选择
-        result = s.pick()
-
-        # CNAllSelector应该返回所有中国市场的股票
-        assert isinstance(result, list)
-        # 具体数量取决于实际的市场数据
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestSelectorIntegration:
-    """测试Selector的集成功能."""
-
-    @pytest.fixture
-    def mock_portfolio(self):
-        """Mock Portfolio实例."""
-        portfolio = Mock(spec=BasePortfolio)
-        portfolio.uuid = "test_portfolio_uuid"
-        return portfolio
-
-    @pytest.fixture
-    def sample_stock_codes(self):
-        """示例股票代码."""
-        return ["000001.SZ", "000002.SZ", "600000.SH", "600594.SH"]
-
-    @pytest.mark.integration
-    @pytest.mark.parametrize("selector_class,init_args", [
-        (FixedSelector, ["test_selector", ["000001.SZ", "000002.SZ"]]),
-        (FixedSelector, ["test_selector_2", ["600000.SH"]]),
-    ])
-    def test_selector_portfolio_integration(self, mock_portfolio, selector_class, init_args):
-        """测试Selector与Portfolio集成."""
-        if BaseSelector is None or FixedSelector is None:
-            pytest.skip("Required selector classes not available")
-
-        s = selector_class(*init_args)
-        s.bind_portfolio(mock_portfolio)
-
-        # 验证绑定成功
-        assert s.portfolio == mock_portfolio
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestSelectorValidation:
-    """测试Selector的验证功能."""
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("invalid_codes", [
-        [],
-        [""],
-        ["INVALID"],
-    ])
-    def test_fixedselector_invalid_codes(self, invalid_codes):
-        """测试FixedSelector无效代码."""
-        if FixedSelector is None:
-            pytest.skip("FixedSelector not available")
-
-        # 尝试创建无效代码的Selector
-        # 具体验证逻辑取决于实现
-        try:
-            s = FixedSelector("test_selector", invalid_codes)
-            # 如果没有验证，那么创建会成功
-            assert s is not None
-        except (ValueError, AttributeError):
-            # 如果有验证，抛出异常是预期的
-            pass
-
-    @pytest.mark.unit
-    @pytest.mark.parametrize("invalid_limit", [
-        0,
-        -100,
-        1000000,  # 过大的值
-    ])
-    def test_popularityselector_invalid_limit(self, invalid_limit):
-        """测试PopularitySelector无效限制."""
-        if PopularitySelector is None:
-            pytest.skip("PopularitySelector not available")
-
-        # 尝试创建无效限制的Selector
-        # 具体验证逻辑取决于实现
-        try:
-            s = PopularitySelector("test_selector", invalid_limit)
-            # 如果没有验证，那么创建会成功
-            assert s is not None
-        except (ValueError, AttributeError):
-            # 如果有验证，抛出异常是预期的
-            pass
-
-
-@pytest.mark.lab
-@pytest.mark.selector
-class TestSelectorComparison:
-    """测试不同Selector的对比."""
-
-    @pytest.fixture
-    def sample_codes(self):
-        """示例代码列表."""
-        return ["000001.SZ", "000002.SZ", "600000.SH"]
-
-    @pytest.mark.unit
-    def test_selector_comparison(self, sample_codes):
-        """对比不同Selector的选择结果."""
-        if FixedSelector is None or CNAllSelector is None:
-            pytest.skip("Required selector classes not available")
-
-        # 创建不同类型的Selector
-        fixed_selector = FixedSelector("fixed", sample_codes[:2])
-        cn_all_selector = CNAllSelector("cn_all")
-
-        # 绑定到相同的portfolio
-        mock_portfolio = Mock(spec=BasePortfolio)
-        fixed_selector.bind_portfolio(mock_portfolio)
-        cn_all_selector.bind_portfolio(mock_portfolio)
-
-        # 执行选择
-        fixed_result = fixed_selector.pick()
-        cn_all_result = cn_all_selector.pick()
-
-        # 验证结果类型
-        assert isinstance(fixed_result, list)
-        assert isinstance(cn_all_result, list)
